gamemaster/pyproject.toml
[project]
name = "monopoly-gamemaster"
version = "0"
requires-python = "==3.14.*"
dependencies = [
  "paho-mqtt",
  "django",
  "django-extensions",
  "djangorestframework ",
  "pydash",
  "drf-standardized-errors ",
  "django-ordered-model",
  "pygltflib ",
  "pytest-django ",
  "django-configurations",
  "python-statemachine",
  "drf-spectacular ",
  "psycopg[binary]",
  "uvicorn",
  "pytest-playwright",
  "servestatic>=3.1.0",
  "django-storages[s3]>=1.14.6",
]
[dependency-groups]
dev = [
    "pytest-django",
    "ruff",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["gamemaster"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
DJANGO_CONFIGURATION = "Testing"
testpaths = ["tests"]
addopts = "--nomigrations --tb=short"
cache-dir = "/tmp/monopoly/ruff_cache"

[tool.ruff]
line-length = 125
cache-dir = "/tmp/monopoly/ruff_cache"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "PT", "ARG001", "ARG002", "RUF012", "ISC", "COM"]

[tool.ruff.lint.isort]
force-single-line = true
lines-after-imports = 2
known-first-party = ["gamemaster", "tests"]
from-first = false
case-sensitive = false
